split feature layer and marker feature ITs
